requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
joblib
matplotlib